# openapi.yaml
openapi: 3.0.3
info:
    title: Notes API (Azure Aligned)
    description: Secure, multi-user, database-backed Notes REST API for IEEE x GitHub Codeathon.
    version: 1.1.0
servers:
    - url: https://notesapi-sgvx.onrender.com
      description: Production server
    - url: /
      description: Dynamic server host (Local)

paths:
    /health:
        get:
            summary: Health check
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
                      example: "2024-05-25"
            responses:
                "200":
                    description: OK

    /auth/register:
        post:
            summary: Register a new user
            tags: [auth]
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RegisterRequest"
            responses:
                "201":
                    description: Created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserResponse"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "409":
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

    /auth/login:
        post:
            summary: Login and get tokens
            tags: [auth]
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TokenResponse"
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

    /auth/refresh:
        post:
            summary: Refresh access token
            tags: [auth]
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RefreshRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TokenResponse"

    /auth/logout:
        post:
            summary: Logout (revoke refresh token)
            tags: [auth]
            security:
                - bearerAuth: []
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RefreshRequest"
            responses:
                "200":
                    description: OK

    /auth/forgot-password:
        post:
            summary: Request password reset
            tags: [auth]
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ForgotPasswordRequest"
            responses:
                "200":
                    description: OK

    /auth/reset-password:
        post:
            summary: Reset password with token
            tags: [auth]
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ResetPasswordRequest"
            responses:
                "200":
                    description: OK

    /notes:
        get:
            summary: List all notes (owned or all if admin)
            tags: [notes]
            security:
                - bearerAuth: []
            parameters:
                - name: api-version
                  in: query
                  required: true
                  schema:
                      type: string
                - name: top
                  in: query
                  schema:
                      type: integer
                - name: skip
                  in: query
                  schema:
                      type: integer
                - name: orderby
                  in: query
                  schema:
                      type: string
                - name: filter
                  in: query
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/NoteListResponse"

        post:
            summary: Create a new note
            tags: [notes]
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/NoteInput"
            responses:
                "201":
                    description: Created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"

    /notes/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            summary: Get a note by ID
            tags: [notes]
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Note"
                "404":
                    $ref: "#/components/responses/NotFound"

    /admin/users:
        get:
            summary: List all users (Admin only)
            tags: [admin]
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/UserResponse"

components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT

    schemas:
        RegisterRequest:
            type: object
            required: [email, password]
            properties:
                email:
                    type: string
                password:
                    type: string
        LoginRequest:
            type: object
            required: [email, password]
            properties:
                email:
                    type: string
                password:
                    type: string
        RefreshRequest:
            type: object
            required: [refreshToken]
            properties:
                refreshToken:
                    type: string
        TokenResponse:
            type: object
            properties:
                accessToken:
                    type: string
                tokenType:
                    type: string
                expiresIn:
                    type: integer
                refreshToken:
                    type: string
        ForgotPasswordRequest:
            type: object
            required: [email]
            properties:
                email:
                    type: string
        ResetPasswordRequest:
            type: object
            required: [resetToken, newPassword]
            properties:
                resetToken:
                    type: string
                newPassword:
                    type: string
        UserResponse:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                email:
                    type: string
                role:
                    type: string
                createdAt:
                    type: string
                    format: date-time

        Note:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                title:
                    type: string
                body:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time

        NoteInput:
            type: object
            required: [title, body]
            properties:
                title:
                    type: string
                body:
                    type: string
                tags:
                    type: array
                    items:
                        type: string

        NoteListResponse:
            type: object
            properties:
                value:
                    type: array
                    items:
                        $ref: "#/components/schemas/Note"
                nextLink:
                    type: string
                    nullable: true

        AzureErrorDetail:
            type: object
            required: [code, message]
            properties:
                code:
                    type: string
                message:
                    type: string

        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    $ref: "#/components/schemas/AzureErrorDetail"

    responses:
        NotFound:
            description: Not Found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ValidationError:
            description: Validation Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"